# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WEBCLERK 3.0 API",
    "version": "3.0.0",
    "description": "API for managing full ERP system with role-based access, email verification, and JWT authentication."
  },
  "paths": {
...
    "/wcapi/save/": {
      "post": {
        "operationId": "wcapi_save_create_update",
        "description": "Create or update a record by model_name. If id is provided, updates that record; otherwise creates a new record. Returns JSON envelope with saved record and messages.",
        "tags": [
          "wcapi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WcapiSaveRequest"
              },
              "examples": {
                "UpdateContact": {
                  "value": {
                    "model_name": "contact",
                    "id": 1,
                    "name_first": "fred",
                    "user1": "test of undefined",
                    "needtoremove": "find a way"
                  },
                  "description": "Update existing contact id=1; unknown fields are ignored or captured in prefs.userdefined"
                }
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/WcapiSaveRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/WcapiSaveRequest"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "jwtAuth": []
          },
          {
            "cookieAuth": []
          },
          {
            "BearerAuth": {
              "type": "http",
              "scheme": "bearer",
              "bearerFormat": "JWT"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WcapiSaveEnvelope"
                },
                "examples": {
                  "SaveResponse": {
                    "value": {
                      "status": "success",
                      "error": null,
                      "code": 200,
                      "message": "",
                      "data": {
                        "id": 1,
                        "model_name": "contact",
                        "version": 2,
                        "record": {
                          "id": 1,
                          "name_first": "fred",
                          "role": "user"
                        },
                        "messages": []
                      }
                    },
                    "description": "Response after save (update or create)"
                  }
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WcapiSaveError"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WcapiSaveAuthError"
                }
              }
            },
            "description": ""
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WcapiSaveVersionConflict"
                }
              }
            },
            "description": ""
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "WcapiSaveAuthError": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      },
      "WcapiSaveEnvelope": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "error": {
            "nullable": true
          },
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/WcapiSaveResponse"
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "status"
        ]
      },
      "WcapiSaveError": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      },
      "WcapiSaveRequest": {
        "type": "object",
        "properties": {
          "model_name": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "version": {
            "type": "integer"
          },
          "expected_version": {
            "type": "integer"
          }
        },
        "required": [
          "model_name"
        ]
      },
      "WcapiSaveResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "model_name": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "nullable": true
          },
          "record": {
            "type": "object",
            "additionalProperties": {}
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "model_name",
          "record"
        ]
      },
      "WcapiSaveVersionConflict": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid"
      },
      "jwtAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}